requirements.txt
torch
torchvision
requests
beautifulsoup4
pdfplumber
Pillow
medmnist
tqdm
